Refactored SequenceExpression building
I really think I nailed it this time!
SequenceExpression now yields a SequenceProduct, which behaves like an array but can be upcast to provide a Tuple-like interface with only a small performance hit. This greatly simplifies how SequenceExpressions are handled, as you no longer have to come up with argument names for unwanted expressions, and matchActions always have a single parameter.
Expanded sequence building functions to support up to 16 expressions.