Replaced CharacterRangeParsingExpression with CharacterSetParsingExpression
Replaced CharacterRangeParsingExpression with CharacterSetParsingExpression, which uses a boolean array as a LUT to match char values. This approach should be much more performant than using a Choice of Ranges (at the cost of memory).
Updated IMatchingContext interface to remove code related to matching character ranges
Updated ExpressionBuilder to use CharacterSetParsingExpression